+++ static/js/browser/observation_chart.tsx
@@ -22,7 +22,7 @@ import axios from "axios";
 import React from "react";
 
 import { DataGroup, DataPoint } from "../chart/base";
-import { drawLineChart } from "../chart/draw";
+import { drawLineChart } from "../chart/draw_line";
 import { formatNumber } from "../i18n/i18n";
 import { Series, StatMetadata } from "../shared/stat_types";
 import { randDomId } from "../shared/util";
