+++ static/js/components/tiles/donut_tile.tsx
@@ -23,7 +23,7 @@ import _ from "lodash";
 import React, { useCallback, useEffect, useRef, useState } from "react";
 
 import { DataGroup, DataPoint } from "../../chart/base";
-import { drawDonutChart } from "../../chart/draw";
+import { drawDonutChart } from "../../chart/draw_donut";
 import { PointApiResponse } from "../../shared/stat_types";
 import { NamedTypedPlace, StatVarSpec } from "../../shared/types";
 import { RankingPoint } from "../../types/ranking_unit_types";
