+++ static/js/ranking/ranking_histogram.tsx
@@ -17,7 +17,7 @@
 import React from "react";
 
 import { DataPoint } from "../chart/base";
-import { drawHistogram } from "../chart/draw";
+import { drawHistogram } from "../chart/draw_histogram";
 import { ASYNC_ELEMENT_HOLDER_CLASS } from "../constants/css_constants";
 import { formatNumber } from "../i18n/i18n";
 import { randDomId } from "../shared/util";
